+++ RELEASE-NOTES.md
@@ -8,6 +8,7 @@
 * #296 ECharts: more Axis properties, multiple Y axes
 * #297 PivotBuilder - align method naming with DataFrame
 * #298 WindowBuilder - support "range" in "agg(..)"
+* #300 GroupBy.map(..) / select(..) - create columns for positions that do not match
 
 ## Release 1.0.0-M21